requirements.txt
# 核心依赖

# 图像处理
Pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0

# Web界面
gradio>=3.50.0

# 工具
tqdm>=4.62.0
pyyaml>=6.0
scikit-learn>=1.0.0